Feature: Stepper settings menu to choose datasource and Scope UUIDs
What? :
A stepper menu in which the user can set the Datasource, choose the scopes
they want as datasources and has default settings to start a measurment.

Why ?: In the current software the user can choose the datasource and has
to start the measurement directly . There is no option to choose
which Scopes they want to connect and everytime the user opens the
software they have to select a source. It is not possible to have default
settings to start and stop the measurement. A settings menu would make this
possible

How ? : A stepper module that pops up via a settings button. The settings are
saved in a settings service. The default Datasource ist dummy data.