Feature: Basic Setup of DevDataServer as new Datasource
This is the minimal needed integration of the DevDataServer as a new
DataSource. The Datasource now shows in the software as such but has no
functionality.

These were the steps taken to integrate the Datasource
1. Integration of the TestdataServer as a Datasource in the availableSources
in the DataSourceSelectionService

Why?
The source needs to be available in the DataSourceSelectionService to be
reachable as a possible datasource

2. Enhancement : Changed the omnai-scope-server folder name to data-server as the
services inside the folder are for different datasources .

3. Setup of a minimal service that implements the DataSource interface
   This contains:
   - a connect function
   - a data signal
   The service only provides the interface implementation not the functionality.