Feature: Button to get devices from testserver
Problem : Without Information about the devices the Dataserver provides
it is not possible to receive data from the server.

Solution: An easy way to implement this in a first iteration without
polling the device info every few seconds, is a simple button that gets
the Device Information once.
Therefor a temporary button: Load Test Server Devices list was
implemented

Userflow:
1. Click on Load Test Server Devices list button
2. Devices should be shown
3. Click on Start data
4. Choose the test data server
5. Click on Start
6. The data should be shown in the graph

Note: This is a simple example for the testdataserver on how this could
be achieved.