Feature: Get Devices from Testserver button
The User needs to be able to get the devices from the Testserver to
start a measurement.
Therefor a temporary button: Load Test Server Devices list was
implemented

Userflow:
1. Click on Load Test Server Devices list button
2. Devices should be shown
3. Click on Start data
4. Choose the test data server
5. Click on Start
6. The data should be shown in the graph

Note: This userflow is furthermore improved by a PR that is currently in
development for the OmnAIScope Dataserver where the user can choose the
UUIDs they want to receive data from and the current devices are
constantly received by polling. To integrate this feature also for the
other datasources a strategy pattern should be designed.