Feature: Select UUIDs from DevDataServer via Popup
Issue: When starting the measurement, all UUIDs were selected by
default. It should be possible to configure which channel the user wants
to receive from the websocket.

Changes: PopupModal that starts from within the connect function of the
dataserver. Therefore the dataselection and handling different
configurations for different dataservers stays separated.

This is a minimal example to show how a configuration menu for a
websocket can be integrated in the software .

Userflow:

- Load Test data via the load Test server devices list button
- Click on start data
- Select Testserver as source
- Click Start
- Configure your Websocket by selecting the prefered UUID
- Click on Connect
You should be able to see the waveform from the sinus or rectangular
generator

Note: This example does not provide a stop or delete possibility
therefore connection is only possible once .